pyproject.toml
[project]
name = "reaper"
version = "0.1.0"
description = ""
authors = [
    { name = "Gecko", email = "gecko.eidechse+git@pm.me" },
    { name = "Cyn", email = "70904206+CooldudePUGS@users.noreply.github.com" },
]
requires-python = "~=3.11"
readme = "README.md"
license = "MIT"
dependencies = [
    "discord-py>=2.5.2",
    "wheel>=0.45.1",
    "beautifulsoup4>=4.13.3",
    "lxml>=5.3.2",
    "pytesseract>=0.3.13",
    "requests>=2.32.3",
    "coloredlogs>=15.0.1",
    "accelerate>=1.6.0",
]

[dependency-groups]
dev = ["ruff>=0.11.5"]
sentiment-analyzer = [
    "transformers>=4.50.3",
    "torch>=2.6.0",
]

[tool.uv]
default-groups = [
    "dev",
    "sentiment-analyzer",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = [
    "A004", # Check for import shadowing builtin functions: https://docs.astral.sh/ruff/rules/builtin-import-shadowing/
    "G010", # Add check for logger.warn
]
extend-ignore = [
    "E402", # Ignore: Module level import not at top of file
]
